I can't figure out how to get the user details like in splash_screen.dart for main.dart.
I want to check if we have a valid email in the user variable, and if so, forward to the chat page.  If not, we should just display the normal welcome page.I can't figure out how to get the user details like in splash_screen.dart for main.dart.

I want to check if we have a valid email in the user variable, and if so, forward to the chat page.  If not, we should just display the normal welcome page.

Signed-off-by: Hans Kokx <hans.d.kokx@gmail.com>